pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-farmbot_advisor"
version = "0.1.0"
description = "Farmbot Advisor - Agricultural decision-making RL environment for Indian farmers"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # Server dependencies
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    # LLM inference
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
server = "farmbot_advisor.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["farmbot_advisor", "farmbot_advisor.server"]
package-dir = { "farmbot_advisor" = ".", "farmbot_advisor.server" = "server" }
